Add hover variant to RelatedWorksCard
